+++ integration-tests/library/src/main/java/com/squareup/anvil/test/PriorityBindings.kt
@@ -1,0 +1,15 @@
+package com.squareup.anvil.test
+
+import com.squareup.anvil.annotations.ContributesBinding
+import com.squareup.anvil.annotations.ContributesBinding.Priority.HIGH
+
+public interface PriorityBinding
+
+@ContributesBinding(
+  scope = AppScope::class,
+  priority = HIGH
+)
+public object PriorityBindingHigh : PriorityBinding
+
+@ContributesBinding(AppScope::class)
+public object PriorityBindingNormal : PriorityBinding